Fix build issue due to spotbug verification error
