+++ websubhub-native/spotbugs-exclude.xml
@@ -18,7 +18,7 @@
 <FindBugsFilter>
     <Match>
         <OR>
-            <Package name="org.ballerinalang.net.websub.hub*"/>
+            <Package name="io.ballerina.stdlib.websubhub.*"/>
         </OR>
         <OR>
             <Bug pattern="IS2_INCONSISTENT_SYNC"/>
